+++ kotlintest-core/src/main/kotlin/io/kotlintest/extensions/Extension.kt
@@ -1,6 +1,11 @@
 package io.kotlintest.extensions
 
 /**
- * Marker interface for all KotlinTest extensions.
+ * Marker interface for extensions that can be added project wide.
  */
-interface Extension
+interface ProjectLevelExtension
+
+/**
+ * Marker interface for extensions that can be added to specs.
+ */
+interface SpecLevelExtension